fix: Update parking_lot and related dependencies to version 0.12.4; refactor KARI references to gas_coin
