+++ Cargo.toml
@@ -204,7 +204,7 @@ num-derive = "0.4.2"
 num-traits = "0.2.19"
 once_cell = "1.21.3"
 ordinals = "0.0.15"
-parking_lot = "0.12.3"
+parking_lot = "0.12.4"
 petgraph = "0.6.5"
 primitive-types = { version = "0.12.1", features = ["serde", "arbitrary"] }
 proptest = "1.7.0"
@@ -365,7 +365,7 @@ tikv-jemallocator = { git = "https://github.com/jamesatomc/jemallocator.git", fe
     "unprefixed_malloc_on_supported_platforms",
     "profiling",
 ] }
-mimalloc = { version = "0.1.46" }
+mimalloc = { version = "0.1.47" }
 multibase = "0.9.0"
 
 # Note: the BEGIN and END comments below are required for external tooling. Do not remove.
